create_Post with search code, landing.php with embedded song
See line 19 of create_Post: link only changes after search?q with the query (How to handle multiple words in search? Could we perhaps change it/have it be a text entry for user to input? How to parse through response in line 33 and 36?)

See lines 68 - 70 for landing.php, embeds song with compact form (automatically generated in spotify developer website. Cn be song or album, compact preferred for smaller size).